requirements.txt
#django==1.6
#south
#python-memcached 
#django-databrowse
#pytz

sphinx
github3.py

ipython
